backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.9
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
dotenv==0.9.9
exceptiongroup==1.3.1
flake8==7.3.0
gunicorn==23.0.0
iniconfig==2.3.0
kombu==5.6.1
mccabe==0.7.0
packaging==25.0
phonenumbers==9.0.19
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
six==1.17.0
sqlparse==0.5.4
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
